+++ apps/voting/types/global.d.ts
@@ -17,20 +17,51 @@ declare namespace NodeJS {
 }
 
 /**
+ * Representation of a url with associated checksum from the election contract.
+ */
+type ChecksumUrl = {
+    /** The URL of the resource */
+    url: string;
+    /** The sha2 checksum */
+    hash: string;
+};
+
+type CandidateResult = {
+    candidate: {
+        url: string;
+        hash: SDK.HexString;
+    };
+    cummulative_votes: number | bigint;
+};
+
+type FrontendElectionConfig = {
+    election_manifest: ChecksumUrl;
+    election_parameters: ChecksumUrl;
+    candidates: ChecksumUrl[];
+    election_description: string;
+    election_start: number;
+    election_end: number;
+    guardians_setup_done: boolean;
+    guardian_keys: number[][];
+    election_result?: CandidateResult[];
+};
+
+/**
  * The configuration built into the application when served from the backend API
  */
-type Config = {
+type FrontendConfig = {
     /** The URL of the node. Must have grpc-web enabled. */
     node: string;
     /** The contract address of the election contract instance used. */
     contractAddress: { index: string; subindex: string };
     /** The concordium network used. */
     network: TargetNetwork;
+    contractConfig?: FrontendElectionConfig;
 };
 
 /** The device name used for election guard */
 declare const DEVICE_NAME: string;
 /** The configuration built into the application when served from the backend API */
-declare const CONFIG: Config;
+declare const CONFIG: FrontendConfig;
 /** The URL for the backend API */
 declare const BACKEND_API: string;